dupe-ai-backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
sqlalchemy
psycopg2-binary
pgvector
python-multipart
Pillow
requests
numpy
torch
transformers
redis
celery[redis]
tenacity
httpx
loguru
pyjwt
cloudinary
